Simplify testing.
This adds render_authorizedkeysfile_paths which does not
require any interaction with OS.  It is then much easier to test it.

The 4 added tests give pretty good coverage of the new code.

Note, I had to comment out the 'dist.create_user' call as
it was actually trying to create a user.